# Hw10 program optimization