--- a/hw09_struct_validator/go.mod
+++ b/hw09_struct_validator/go.mod
@@ -1,6 +1,7 @@
 module github.com/DimVlas/otus_hw/hw09_struct_validator
 
 go 1.22
+
 require github.com/stretchr/testify v1.9.0
 
 require (