--- a/hw09_struct_validator/rules/rule.go
+++ b/hw09_struct_validator/rules/rule.go
@@ -23,7 +23,7 @@ type FieldRules struct {
 
 // парсит полученную строку, возвращая массив структур с описанием правил проверки
 // ожидается, что строка имеет вид 'правило:условие|правило:условие|...'
-func ParseRulesTag(rulesTag string) ([]RuleInfo, error) {
+func parseRulesTag(rulesTag string) ([]RuleInfo, error) {
 	rulesTag = strings.Trim(rulesTag, " ")
 	if rulesTag == "" {
 		return []RuleInfo{}, nil
@@ -58,19 +58,12 @@ func FieldRulesByTag(fieldName string, fieldTag string) (FieldRules, error) {
 		return frs, nil
 	}
 
-	rs := strings.Split(fieldTag, "|")
-
-	if len(rs) < 1 {
-		return frs, ErrEmptyRule
+	rls, err := parseRulesTag(fieldTag)
+	if err != nil {
+		return frs, err
 	}
-	for _, r := range rs {
-		rule := strings.Split(r, ":")
-		if len(rule) < 2 {
-			return frs, ErrUnknowRule
-		}
 
-		frs.Rules = append(frs.Rules, RuleInfo{Name: rule[0], Cond: rule[1]})
-	}
+	frs.Rules = rls
 
 	return frs, nil
 }
@@ -78,10 +71,10 @@ func FieldRulesByTag(fieldName string, fieldTag string) (FieldRules, error) {
 // типы данных на которые распространяются правила проверки
 //type TypeValue uint
 
-var Rules = map[reflect.Kind]map[string]func(v reflect.Value, condition string) error{
+var Rules = map[reflect.Kind]map[string]func(fName string, v reflect.Value, condition string) error{
 	reflect.String: {
 		// 'len:32' - проверка длины строки должна быть 32 символа
-		"len": func(v reflect.Value, condition string) error {
+		"len": func(fName string, v reflect.Value, condition string) error {
 			if v.Kind() != reflect.String {
 				// это правило применимо только к строкам
 				return fmt.Errorf("this rule applies only to the string")
@@ -94,13 +87,13 @@ var Rules = map[reflect.Kind]map[string]func(v reflect.Value, condition string) 
 
 			if utf8.RuneCountInString(v.String()) != c {
 				return ValidationError{
-					Field: "",
+					Field: fName,
 					Err:   fmt.Errorf("length of the string not equal to %s", condition),
 				}
 			}
 			return nil
 		},
-		"regexp": func(v reflect.Value, condition string) error {
+		"regexp": func(fName string, v reflect.Value, condition string) error {
 			if v.Kind() != reflect.String {
 				return fmt.Errorf("this rule applies only to the string")
 			}
@@ -111,70 +104,15 @@ var Rules = map[reflect.Kind]map[string]func(v reflect.Value, condition string) 
 			}
 			if !pattern.MatchString(v.String()) {
 				return ValidationError{
-					Field: "",
+					Field: fName,
 					Err:   fmt.Errorf("length of the string not equal to %s", condition),
 				}
 			}
 
 			return ErrRuleNotImplement
 		},
-		"in": func(v reflect.Value, condition string) error {
+		"in": func(fName string, v reflect.Value, condition string) error {
 			return ErrRuleNotImplement
 		},
 	},
 }
-
-// const (
-// 	Int TypeValue = iota
-// 	IntSlice
-// 	String
-// 	StringSlice
-// )
-
-// type Rule interface {
-// 	Rule() string
-// 	Validate(reflect.Value) error
-// }
-
-// func New(rule string) (Rule, error) {
-// 	if len(strings.Trim(rule, " ")) == 0 {
-// 		return nil, ErrEmptyRule
-// 	}
-
-// 	tag := reflect.StructTag(rule)
-// 	r, ok := tag.Lookup("len")
-// 	if !ok {
-// 		return nil, errors.New(" not len rule")
-// 	}
-
-// 	l, err := strconv.Atoi(r)
-// 	if err != nil {
-// 		return nil, fmt.Errorf("can't parse the '%s' rule", rule)
-// 	}
-// 	lenS := LenStr{
-// 		rule: rule,
-// 		len:  l,
-// 	}
-// 	return lenS, nil
-// }
-
-// type LenStr struct {
-// 	rule string
-// 	len  int
-// }
-
-// func (r LenStr) Rule() string {
-// 	return r.rule
-// }
-// func (r LenStr) Validate(val reflect.Value) error {
-// 	if val.Kind() == reflect.String {
-// 		s := val.String()
-
-// 		if len(s) != r.len {
-// 			return fmt.Errorf("the length should be equal to %d", r.len)
-// 		}
-// 		return nil
-// 	}
-
-// 	return fmt.Errorf("the '%s' rule only applies to values of type'%s'", r.rule, "string")
-// }