--- a/hw09_struct_validator/rules/validate_test.go
+++ b/hw09_struct_validator/rules/validate_test.go
@@ -1,1 +1,0 @@
-package rules