… .golangci.yml
… hw09_struct_validator/rules/RULES.md
… hw09_struct_validator/rules/errors_test.go
… hw09_struct_validator/rules/rule_test.go
… hw09_struct_validator/rules/rules_test.go
… hw09_struct_validator/rules/validate.go
… hw09_struct_validator/validate.go
… hw09_struct_validator/validator.go
… hw09_struct_validator/validator_test.go
… hw10_program_optimization/bench_new.txt
… hw10_program_optimization/bench_old.txt
… hw10_program_optimization/bench_stat.txt
… hw10_program_optimization/go.mod
… hw10_program_optimization/go.sum
… hw10_program_optimization/stats.go
… hw10_program_optimization/stats_easyjson.go
… hw10_program_optimization/stats_test.go